fyp/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
chardet==3.0.4
Django==3.2.5
django-cors-headers==3.7.0
django-extensions==3.1.5
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
idna==2.8
portpicker==1.2.0
PyJWT==2.3.0
pytz==2021.3
requests==2.21.0
simplegeneric==0.8.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.24.3
wincertstore==0.2
